requirements.txt
pymongo[srv]
praw
python-dotenv
pytest
pytest-mock